requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
